requirements.txt
PyInputPlus==0.2.12
PyYAML==6.0
rich==12.4.4
